Refactor project structure and update configuration
- Move audioSegmentation.py to modules/audio/ directory for better organization
- Update import statements in trackRecognition and trackValidation modules
- Rename log directory to logs in constants.py for consistency
- Update .gitignore to include all directories from constants.py
- Revise project structure documentation in README.md
- Maintain backward compatibility while improving code organization
