+++ modules/core/constants.py
@@ -1,0 +1,23 @@
+# Application Title
+APP_TITLE = 'Shazam Tool'
+
+# Directory for application logs
+APPLICATION_LOGS_DIR = 'logs'
+
+# Directory for downloaded mp3 files
+DOWNLOADS_DIR = 'downloads'
+
+# Directory for result txt files
+RESULTS_DIR = 'recognised-lists'
+
+# Directory for audio segments
+AUDIO_SEGMENTS_DIR = 'audio-segments'
+
+# Duration of each segment in milliseconds (10 seconds)
+SEGMENT_LENGTH = 10 * 1000
+
+# Maximum number of retries for Shazam API requests
+MAX_RETRIES = 3
+
+# Timeout for Shazam API requests
+TIMEOUT = 40
